fix becai minimumBaseDefenceFleetSize calculation
there were only integer divisions which is probably not what was intended here (the casts to double were missing)
found by msvc complaining about the unknown std::ceil(int) overload
(cherry picked from commit d01e9c3fe8977f6f8cd854910db4254f3aafd705)

+++ players/ai/becai/becai.cpp
@@ -205,7 +205,7 @@ AiBecai::play()
     // including us). Actually it is the production that counts and not the
     // planets, but planets are good enough for us here for the moment.
 
-    int minimumBaseDefenceFleetSize = (int) std::ceil(((totalOwnFleet / 2) / totalOwnPlanets) * (1 - (totalOwnPlanets / (totalOwnPlanets + totalEnemyPlanets))));
+    int minimumBaseDefenceFleetSize = (int) std::ceil(((totalOwnFleet / 2.0) / totalOwnPlanets) * (1 - ((double) totalOwnPlanets / (totalOwnPlanets + totalEnemyPlanets))));
     int originalMinimumBaseDefenceFleetSize = minimumBaseDefenceFleetSize;
 
     // Try to keep the defence above the universe average so that our planets
